requirements.txt
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-generativeai
sentence-transformers
faiss-cpu
python-dotenv
numpy
pymupdf